+++ src/qibocal/protocols/ramsey/__init__.py
@@ -1,5 +1,5 @@
-from .ramsey import ramsey
-from .ramsey_signal import ramsey_signal
-from .ramsey_zz import ramsey_zz
+from .classification import ramsey
+from .signal import ramsey_signal
+from .zz_coupling import ramsey_zz
 
 __all__ = ["ramsey", "ramsey_signal", "ramsey_zz"]
