+++ src/qibocal/protocols/ramsey/signal.py
@@ -11,14 +11,14 @@ from qibocal.config import log
 from qibocal.protocols.utils import table_dict, table_html
 from qibocal.result import unpack
 
-from .ramsey_acquisition import (
+from .acquisition import (
     RamseyData,
     RamseyParameters,
     RamseyResults,
     execute_experiment,
     ramsey_sequence,
 )
-from .utils import fitting, process_fit, ramsey_fit, ramsey_update
+from .processing import fitting, process_fit, ramsey_fit, ramsey_update
 
 __all__ = ["ramsey_signal"]
